Fix stack alignment for JIT function call targets
X64_84 ABIs demand that the stack pointer is a multiple of 16 before
CALLing a function. Compilers such as GCC create optimized code
based on this constraint, e.g. code with MOVAPS in which the operand
must be aligned on a 16-byte boundary.

The dynamips JIT does not ensure alignment before CALLs. This led to
SIGSEGV crashes of dynamips in libc on Ubuntu 22.04 and earlier.

There are multiple options to deal with this:
    1. Fix in the caller: The JIT is changed to generate code that
                          ensures proper stack pointer alignment
    2. Fix in the callee: Functions called from JIT generated code
                          ensure proper stack pointer alignment

Option 1 is implemented in this change by adding two macros that
are invoked around JIT generated CALL sites to align the stack
pointer to a 16-byte boundary.

Two variants are implemented (fast and universal) of which one can
be chosen using either
	cmake -DAMD64_STACK_ALIGN_CALL=fast		or
	cmake -DAMD64_STACK_ALIGN_CALL=universal

The "fast" variant assumes that the stack pointer is off by 8,
which might not always be true but worked during tests. In case
this fails the "universal" variant can be used.

This whole solution stops to work once the called functions accept
parameters passed via the stack. That is because the realignment
changes the offsets to parameters.